Split out methods into own files
