+++ imports/startup/both/methods.js
@@ -1,4 +1,5 @@
 import '/imports/api/users/methods';
 import '/imports/api/interactions/methods';
+import '/imports/api/interactions/participationVotingMethods';
 import '/imports/api/submissions/methods';
 import '/imports/api/menu/methods';
